Added sanity checks to add-remote-machine.bash.

+++ add-remote-machine.bash
@@ -12,24 +12,73 @@
 
 set -eu
 
-# add-remote-machine.bash juju-ci3 ./staging-juju-rsa ubuntu@10.245.67.134
-env=$1
-ssh_key=$2
-user_at_host=$3
+check_access() {
+    name=$1
+    url=$2
+    echo "Checking $USER_AT_HOST can access $name at $url..."
+    set +e
+    result=$(ssh -i $SSH_KEY $USER_AT_HOST curl --connect-timeout 2 --head $url || echo "fail")
+    set -e
+    result=$(tail -1 $result)
+    echo "== $result"
+    if [[ $result == "fail" ]]; then
+        echo "...FAIL"
+        NETWORK_ACCESS="false"
+    else
+        echo "...OK"
+    fi
+}
 
-# Learn the state-server's public IP and private internl name DNS name.
-private_dns_name=$(juju ssh 0 'echo "$(hostname).$(dnsdomainname)"')
-public_dns_name=$(juju status | sed -r '/dns-name/!d; 1,1!d; s/.*: (.*)/\1/')
+
+check_url_access() {
+    option=$1
+    option_url=$(juju get-env -e $ENV $option)
+    if [[ -n "$option_url" ]]; then
+        check_access $option $option_url
+    else
+        echo "...! You must verify that $USER_AT_HOST can access $option"
+    fi
+}
+
+
+if [[ "$1" == "--dry-run" ]]; then
+    DRY_RUN="true"
+    shift
+else
+    DRY_RUN="false"
+fi
+ENV=$1
+SSH_KEY=$2
+USER_AT_HOST=$3
+NETWORK_ACCESS="true"
+
+private_dns_name=$(juju ssh -e $ENV 0 'echo "$(hostname).$(dnsdomainname)"')
+public_dns_name=$(juju status -e $ENV | sed -r '/dns-name/!d; 1,1!d; s/.*: (.*)/\1/')
 public_ip=$(dig "$public_dns_name" | sed -r '/^;/d; /IN/!d; s/.*A (.*)/\1/;')
-
+api_port=$(juju get-env -e $ENV api-port)
 echo "State-server public address: $public_ip $public_dns_name"
 echo "State-server private address: $private_dns_name"
+check_access "state-server" http://$public_ip:$api_port
+check_url_access image-metadata-url
+check_url_access tools-metadata-url
 
-# Add the state-server internal ip and address to the gateway's /etc/hosts
-echo "Adding state-server's private dns name to $user_at_host:/etc/hosts"
-ssh -i $ssh_key $user_at_host \
+echo "Checking $USER_AT_HOST can access the cloud provider's storage"
+provider=$(juju get-env -e $ENV "type")
+if [[ $provider == "ec2" ]]; then
+    control_bucket=$(juju get-env -e $ENV control-bucket)
+    check_access "s3" http://s3ss.amazon.com/$control_bucket
+else
+    echo "...! You must verify that $USER_AT_HOST can access the cloud storage."
+fi
+
+
+if [[ $DRY_RUN == "true" || $NETWORK_ACCESS == 'false' ]]; then
+    exit
+fi
+
+echo "Adding state-server's private dns name to $USER_AT_HOST:/etc/hosts"
+ssh -i $SSH_KEY $USER_AT_HOST \
     "echo '$public_ip $private_dns_name' | sudo tee -a /etc/hosts"
 
-# Add the machine to juju-ci using the private address
-echo "Adding $user_at_host to $env"
-juju --show-log add-machine ssh:$user_at_host
+echo "Adding $USER_AT_HOST to $ENV"
+juju --show-log add-machine -e $ENV ssh:$USER_AT_HOST
